fix: code organisation and rendering
