+++ src/components/Search/command.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Clipboard } from 'lucide-react';
 import { DiceCmdMeta } from '@/data/command';
 import { useState } from 'react';
@@ -14,11 +16,11 @@ export default function CommandPage({
   onCopy,
 }: CommandPageProps) {
   const [isCopied, setIsCopied] = useState(false);
-  setIsCopied(true);
-  setTimeout(() => {
-    setIsCopied(false);
-  }, 1000);
   const handleCopy = () => {
+    setIsCopied(true);
+    setTimeout(() => {
+      setIsCopied(false);
+    }, 1000);
     navigator.clipboard.writeText(syntax).then(() => {
       if (onCopy) {
         onCopy();
